# openapi.yaml
# DO NOT EDIT
# Auto generated by "flask gen"

info:
  description: What have you done today?
  title: Timetable
  version: 0.1.0
paths:
  /api/ping:
    get:
      summary: Ping API server.
      tags:
      - common
      x-swagger-router-controller: timetable.views
      operationId: ping
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: pong
  /api/user/login:
    post:
      summary: User login
      tags:
      - user
      x-swagger-router-controller: timetable.views.user.index
      operationId: user_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginForm'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
  /api/event/categories:
    get:
      summary: Get event categories.
      tags:
      - event
      x-swagger-router-controller: timetable.views.event
      operationId: get_event_categories
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'
  /api/event/save:
    post:
      summary: Save event.
      tags:
      - event
      x-swagger-router-controller: timetable.views.event
      operationId: save_event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
  /api/event/list:
    get:
      summary: Get event list.
      tags:
      - event
      x-swagger-router-controller: timetable.views.event
      operationId: get_event_list
      parameters:
      - in: query
        name: start
        schema:
          type: string
          format: date-time
      - in: query
        name: end
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
  /api/event/delete:
    post:
      summary: Delete event.
      tags:
      - event
      x-swagger-router-controller: timetable.views.event
      operationId: delete_event
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
  /api/note/content:
    get:
      summary: Get note content.
      tags:
      - note
      x-swagger-router-controller: timetable.views.note
      operationId: get_note_content
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteForm'
  /api/note/save:
    post:
      summary: Save note.
      tags:
      - note
      x-swagger-router-controller: timetable.views.note
      operationId: save_note
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NoteForm'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
openapi: 3.0.2
components:
  schemas:
    LoginForm:
      type: object
      properties:
        password:
          type: string
        username:
          type: string
      required:
      - password
      - username
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
    Category:
      type: object
      properties:
        title:
          type: string
        id:
          type: integer
        color:
          type: string
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          minLength: 1
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        categoryId:
          type: integer
      required:
      - categoryId
      - end
      - start
      - title
    NoteForm:
      type: object
      properties:
        content:
          type: string
    Note:
      type: object
      properties:
        created:
          type: string
          format: date-time
        content:
          type: string